backend/requirements.txt
aiofiles==0.8.0
aniso8601==9.0.1
asgiref==3.5.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.52.0
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.5
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
passlib==1.7.4
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-multipart==0.0.4
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
starlette==0.13.2
typing_extensions==4.1.1
ujson==5.1.0
urllib3==1.26.8
uvicorn==0.17.5
zipp==3.7.0
